clean up logging work
