+++ docs/logs/extending-the-sdk/MyExporter.cs
