+++ docs/logs/extending-the-sdk/MyProcessor.cs
