+++ docs/logs/extending-the-sdk/README.md
@@ -1,0 +1,20 @@
+# Extending the OpenTelemetry .NET SDK
+
+* [Building your own exporter](#exporter)
+* [Building your own processor](#processor)
+* [Building your own sampler](#sampler)
+* [References](#references)
+
+## Exporter
+
+TBD
+
+## Processor
+
+TBD
+
+## Sampler
+
+TBD
+
+## References
